MLX implementation

+++ src/models/face_detect.py
@@ -64,10 +64,10 @@ def detect_faces(image_bytes: bytes) -> tuple[list[dict], int, int]:
         
         face_data = {
             "boundingBox": {
-                "x1": float(x1),
-                "y1": float(y1),
-                "x2": float(x2),
-                "y2": float(y2)
+                "x1": int(x1),
+                "y1": int(y1),
+                "x2": int(x2),
+                "y2": int(y2)
             },
             "score": float(observation.confidence()),
         }
